requirements.txt
# jax==0.6.2
# jaxlib==0.6.2 (with cuda or tpu support, see https://github.com/jax-ml/jax#installation)
# tensorflow==2.16.1
numpyro==0.14.0
flax==0.10.7
tqdm==4.66.1
hydra-core==1.3.2
hydra-joblib-launcher==1.2.0
hydra-submitit-launcher==1.2.0
# wandb==0.15.10
ml-collections==0.1.1
ott-jax==0.4.6
tensorflow-probability==0.24.0
dm-haiku==0.0.12
distrax==0.1.5
mergedeep==1.3.4
matplotlib==3.8.4
pandas==2.1.0
comet_ml==3.55.0
